Changed package name from com.proofmode to org.proofmode
