--- app/src/main/java/com/proofmode/c2pa/data/Media.kt
@@ -1,9 +1,0 @@
-package com.proofmode.c2pa.data
-
-import android.net.Uri
-
-data class Media(
-    val uri: Uri,
-    val isVideo: Boolean,
-    val date: Long,
-)
