+++ app/src/main/java/org/proofmode/c2pa/C2PAApp.kt
@@ -1,0 +1,16 @@
+package org.proofmode.c2pa
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+import timber.log.Timber
+
+@HiltAndroidApp
+class C2PAApp: Application() {
+    override fun onCreate() {
+        super.onCreate()
+        if (BuildConfig.DEBUG) {
+            Timber.plant(Timber.DebugTree())
+        }
+
+    }
+}
