+++ app/src/main/java/org/proofmode/c2pa/c2pa_signing/selfsign/CertificateSigningService.kt
