+++ app/src/main/java/org/proofmode/c2pa/data/Media.kt
@@ -1,0 +1,9 @@
+package org.proofmode.c2pa.data
+
+import android.net.Uri
+
+data class Media(
+    val uri: Uri,
+    val isVideo: Boolean,
+    val date: Long,
+)
