+++ app/src/main/java/org/proofmode/c2pa/ui/RecordingState.kt
@@ -1,0 +1,11 @@
+package org.proofmode.c2pa.ui
+
+import android.net.Uri
+
+sealed class RecordingState {
+    object Idle : RecordingState()
+    object Recording : RecordingState()
+    object Paused : RecordingState()
+    data class Finalized(val outputUri: Uri?) : RecordingState()
+    data class Error(val message: String?) : RecordingState()
+}
